# api/swagger.yaml
basePath: /api/v1
definitions:
  handlers.ColorChangeResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  handlers.ColorsListResponse:
    properties:
      supportedColors:
        items:
          type: string
        type: array
    type: object
  handlers.LightStateChangeResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  handlers.SoundPlayResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  handlers.SoundsListResponse:
    properties:
      supportedSounds:
        items:
          type: string
        type: array
    type: object
host: witchonstephendrive.com
info:
  contact:
    email: caleblemoine@gmail.com
    name: Caleb Lemoine
  description: Control my halloween decorations
  license:
    name: MIT
    url: https://raw.githubusercontent.com/circa10a/witchonstephendrive.com/main/LICENSE
  title: witchonstephendrive.com
  version: 0.1.0
paths:
  /color/{color}:
    post:
      description: Change hue lights to supported color defined in color map
      parameters:
      - description: Color to change lights to
        in: path
        name: color
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ColorChangeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ColorChangeResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ColorChangeResponse'
      summary: Change hue lights color
  /colors:
    get:
      description: Get list of supported colors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ColorsListResponse'
      summary: Get available colors to change to
  /lights/{state}:
    post:
      description: Only supports states of on/off
      parameters:
      - description: State to set lights to (on/off)
        in: path
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LightStateChangeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.LightStateChangeResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.LightStateChangeResponse'
      summary: Change state of configured lights
  /sound/{sound}:
    post:
      description: Play halloween sound supported in sound list
      parameters:
      - description: Sound to play
        in: path
        name: sound
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handlers.SoundPlayResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.SoundPlayResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handlers.SoundPlayResponse'
      summary: Play sound via assistant relay
  /sounds:
    get:
      description: Get list of supported sounds
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SoundsListResponse'
      summary: Get available sounds to play
schemes:
- https
swagger: "2.0"